requirements.txt
strands-agents
strands-agents[a2a]
strands-agents-tools
mcp